Updated README.md
Updated the README file to better represent the actual userChrome.css file which is distributed in the releases section. Most of the values are switched around and the default values for `--menu-item-height: 35px;` and `--menu-corner-rounding: 11px;` were changed from 30 to 35 and 10 to 11 respectively, following the defaults given in the latest release. The `--toolbarbutton-hover-background: rgba(127, 127, 127, 0.5);` value has (still) been left out as it isn't touched in any of the examples. However, if someone were willing to redo the GIFs/examples, this value could be showcased as well.

The goal here is to make it a bit easier for new users to change the correct values in their userChrome.css file.